Move GameDetails into stats types
