+++ src/containers/FileListInput.tsx
@@ -5,7 +5,8 @@ import { DropPad } from "components/DropPad";
 import { ErrorMessage } from "components/ErrorMessage";
 import { FileList } from "components/FileList";
 import { processStats } from "lib/processStats";
-import { GameDetails, readFileAsGameDetails } from "lib/readFile";
+import { readFileAsGameDetails } from "lib/readFile";
+import { GameDetails } from "lib/stats/types";
 import { generateSearchParams } from "lib/searchParams";
 import React, { useCallback, useContext } from "react";
 import { useHistory } from "react-router-dom";
