+++ src/lib/stats/types.ts
@@ -1,3 +1,5 @@
+import { FrameEntryType, FramesType, GameEndType, GameStartType, MetadataType, StatsType } from "@slippi/slippi-js";
+
 export enum Stat {
   OPENINGS_PER_KILL = "openingsPerKill",
   DAMAGE_PER_OPENING = "damagePerOpening",
@@ -11,4 +13,14 @@ export enum Stat {
   AVG_KILL_PERCENT = "avgKillPercent",
   HIGH_DAMAGE_PUNISHES = "highDamagePunishes",
   DAMAGE_DONE = "damageDone",
+}
+
+export interface GameDetails {
+  filePath: string;
+  settings: GameStartType;
+  frames: FramesType;
+  stats: StatsType;
+  metadata: MetadataType;
+  latestFrame: FrameEntryType | null;
+  gameEnd: GameEndType | null;
 }
